Add the rest of graphs
Each file is in the style: n-numGraphs_run.json

Training graphs were created in sets of 5000 (though not all sizes of n may have that many solved) and 30 seconds to solve.

Validation graphs were created in sets of 1000 (though not all sizes of n may have that many solved) and either 10 seconds (run 1) or 20 seconds (run 2) to find a solution.